move bind-mount logic to cmd/tether
